Move `HasRateLimit*` classes to server module
